requirements.txt
streamlit
streamlit-chat
langchain
openai
streamlit-option-menu
langchain-experimental
python-dotenv
mysql-connector-python
SQLAlchemy
streamlit-server-state
psycopg2-binary
regex
PyPDF2
openai
faiss-cpu
altair
tiktoken
azure-ai-formrecognizer
pillow
python-dotenv
sqlalchemy
pandas 
numpy
